requirements.txt
Django==4.2.2
djangorestframework==3.14.0
celery==5.4.0
redis==4.5.4
